lieer: ability to override address
The address used to pull a GMail account is assumed to be the one
specified for the account's use.  However, someone might want to
use another address.

Additionally, service accounts set up as delegates can retrieve emails
with "me" set for the user ID.  This is convenient for those of us
who like to keep some references to private email addresses that
we use with proxies off out of publicly published machine configuration.